JM-6244: Updates to PR comments
