+++ server/objects/administration.js
@@ -83,6 +83,52 @@
     },
   };
 
+  module.exports.transportRates = {
+    get: function(app, req, locCode, etag = null) {
+      const payload = {
+        uri: urljoin(config.apiEndpoint, 'moj/court-location', locCode, 'rates'),
+        method: 'GET',
+        headers: {
+          'User-Agent': 'Request-Promise',
+          'Content-Type': 'application/vnd.api+json',
+          Authorization: req.session.authToken,
+        },
+        json: true,
+      };
+
+      if (etag) {
+        payload.headers['If-None-Match'] = `${etag}`;
+      }
+
+      app.logger.info('Sending request to API: ', payload);
+
+      payload.transform = (response, incomingRequest) => {
+        const headers = _.cloneDeep(incomingRequest.headers);
+
+        return { response, headers };
+      };
+
+      return rp(payload);
+    },
+    put: function(app, req, loc, body) {
+      const payload = {
+        uri: urljoin(config.apiEndpoint, 'moj/administration/courts/', loc, 'rates'),
+        method: 'PUT',
+        headers: {
+          'User-Agent': 'Request-Promise',
+          'Content-Type': 'application/vnd.api+json',
+          Authorization: req.session.authToken,
+        },
+        json: true,
+        body,
+      };
+
+      app.logger.info('Sending request to API: ', payload);
+
+      return rp(payload);
+    },
+  };
+
   module.exports.courtsDAO = {
     get: function(app, req) {
       const payload = {
@@ -94,23 +140,6 @@
           Authorization: req.session.authToken,
         },
         json: true,
-      };
-
-      app.logger.info('Sending request to API: ', payload);
-
-      return rp(payload);
-    },
-    putRates: function(app, req, loc, body) {
-      const payload = {
-        uri: urljoin(config.apiEndpoint, 'moj/administration/courts/', loc, 'rates'),
-        method: 'PUT',
-        headers: {
-          'User-Agent': 'Request-Promise',
-          'Content-Type': 'application/vnd.api+json',
-          Authorization: req.session.authToken,
-        },
-        json: true,
-        body,
       };
 
       app.logger.info('Sending request to API: ', payload);
